Cleanup wrappers/readelf.py
