--- tests/binaries/host/.gitignore
@@ -1,1 +1,0 @@
-reference-binary
